Made a change to the clsUMCCreator to include the additional option of using charge state when clustering. The usage of clustering or otherwise is a boolean parameter at the end of the SetOptions and SetOptionsEx methods.
If the charge state is used for Clustering, the CreateUMCsSinglyLinkedWithAll method checks the charge states of the two peaks and decides whether they need to be present in the same cluster.

CAUTION: One might have to rebuild the program and make sure that the VBP parts are seeing the latest method signatures.